Fixes and some error checking
Make sure the experiment paths are below the cwd.
To stop people deleting /

Manually set LD_LIBRARY_PATH, as thttpd does not pass it down to child processes.

Seems to work, but very unreliable at capturing images on the PI.
FLIRA65-Capture was hanging indefinitely without getting any valid frames.
Set FLIRA65-Capture to give up after 5 attempts, and run.py will just keep trying.

+++ manage.cgi
@@ -57,6 +57,13 @@ def parse_form(form):
 			shutil.make_archive("download", "zip", root_dir=".", base_dir=k)
 		elif form.getvalue(k) == 'Delete':
 			if k+"_confirm_delete" in form.keys() and form.getvalue(k+"_confirm_delete") == "Yes":
+				# Only delete paths lower than the current wd
+				root = os.getcwd()
+				k = os.path.abspath(k)
+				if k[0:len(root)] != root or k.count("/") <= root.count("/"):
+					print(bold(red("<p> Path to delete not below root!</p>")))
+					return
+
 				print(bold("<p>Delete %s CONFIRMED</p>" % k))
 				shutil.rmtree(k)
 			else:
